+++ CHANGELOG.md
@@ -5,16 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.0] - 2026-03-30
 
 ### Added
 
 - Added `__str__`, `__repr__` and `__iter__` to **BitAware** [#3](https://github.com/wnowicki/bitaware/issues/3)
+- Added defined flags support to **BitAware**
 
 ### Changed
 
 - Publish action - removed test part
 
+### Removed
+
+- Removed `uv.lock` lock file (locked installs are no longer supported).
+
 ## 0.1.1 - 2025-05-16
 
 [Unreleased]: https://github.com/wnowicki/bitaware/compare/0.1.0...HEAD
+[0.2.0]: https://github.com/wnowicki/bitaware/compare/0.1.0...0.2.0